+++ packages/apptive_grid_form/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.15.0
 * Update apptive_grid_core to 0.15.0
+* Add `onSavedToPending` callback to `ApptiveGridForm` and `ApptiveGridFormData` similar to the `onSuccess` callback to be notified when a Form was submitted and is saved to pending
 
 ## 0.14.4+1
 * Fix Bug where `ApptiveGridFormData` would rebuild when in an "end" state